+++ clients/apple/iOS/Screens/SharedWithMeView.swift
@@ -23,7 +23,7 @@ struct SharedWithMeView: View {
 
     var body: some View {
         Group {
-            if let pendingShares = filesModel.pendingShares {
+            if let pendingShares = filesModel.pendingSharesByUsername {
                 ScrollViewReader { scrollHelper in
                     ScrollView {
                         VStack {
