+++ clients/apple/iOS/Widgets/FileTreeView.swift
@@ -191,81 +191,6 @@ struct FileRowView: View {
     }
 }
 
-struct FileRowContextMenu: View {
-    let file: File
-    
-    @EnvironmentObject var fileTreeModel: FileTreeViewModel
-    @EnvironmentObject var filesModel: FilesViewModel
-    @EnvironmentObject var homeState: HomeState
-    @EnvironmentObject var workspaceInput: WorkspaceInputState
-    
-    var body: some View {
-        VStack {
-            if file.isFolder {
-                Button(action: {
-                    workspaceInput.createDocAt(parent: file.id, drawing: false)
-                }) {
-                    Label("Create a document", systemImage: "doc.fill")
-                }
-                Button(action: {
-                    workspaceInput.createDocAt(parent: file.id, drawing: true)
-                }) {
-                    Label("Create a drawing", systemImage: "pencil.tip.crop.circle.badge.plus")
-                }
-                Button(action: {
-                    homeState.sheetInfo = .createFolder(parent: file)
-                }) {
-                    Label("Create a folder", systemImage: "folder.fill")
-                }
-            }
-            
-            if !file.isRoot {
-                Button(action: {
-                    homeState.sheetInfo = .rename(file: file)
-                }) {
-                    Label("Rename", systemImage: "pencil.circle.fill")
-                }
-
-                Button(action: {
-                    homeState.selectSheetInfo = .move(files: [file])
-                }) {
-                    Label("Move", systemImage: "arrow.up.and.down.and.arrow.left.and.right")
-                }
-                
-                Divider()
-                
-                Button(action: {
-                    homeState.sheetInfo = .share(file: file)
-                }) {
-                    Label("Share", systemImage: "person.wave.2.fill")
-                }
-
-                Button(action: {
-                    exportFiles(homeState: homeState, files: [file])
-                }) {
-                    Label("Share externally to...", systemImage: "square.and.arrow.up.fill")
-                }
-                
-                if file.type == .document {
-                    Button(action: {
-                        ClipboardHelper.copyFileLink(file.id)
-                    }) {
-                        Label("Copy file link", systemImage: "link")
-                    }
-                }
-                
-                Divider()
-                
-                Button(role: .destructive, action: {
-                    filesModel.deleteFileConfirmation = [file]
-                }) {
-                    Label("Delete", systemImage: "trash.fill")
-                }
-            }
-        }
-    }
-}
-
 #Preview {
     NavigationView {
         FileTreeView(root: (AppState.lb as! MockLb).file0, filesModel: .preview, workspaceInput: .preview, workspaceOutput: .preview)
