+++ clients/apple/macOS/Screens/SharedWithMeView.swift
@@ -27,7 +27,7 @@ struct SharedWithMeView: View {
 
     var body: some View {
         Group {
-            if let pendingShares = filesModel.pendingShares {
+            if let pendingShares = filesModel.pendingSharesByUsername {
                 ScrollViewReader { scrollHelper in
                     ScrollView {
                         VStack {
